refactor : 주문 서비스 단일책임 원칙, 개방폐쇄원칙 준수
