+++ src/main/java/com/sparta/uglymarket/order/service/OrderService.java
@@ -13,6 +13,7 @@ import com.sparta.uglymarket.product.repository.ProductRepository;
 import com.sparta.uglymarket.user.entity.User;
 import com.sparta.uglymarket.user.repository.UserRepository;
 import com.sparta.uglymarket.util.JwtUtil;
+import com.sparta.uglymarket.util.UserFinder;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -22,13 +23,13 @@ import java.util.List;
 public class OrderService {
 
     private final OrderRepository orderRepository;
-    private final UserRepository userRepository;
+    private final UserFinder userFinder;
     private final ProductRepository productRepository;
 
 
-    public OrderService(OrderRepository orderRepository, UserRepository userRepository, ProductRepository productRepository) {
+    public OrderService(OrderRepository orderRepository, UserFinder userFinder, ProductRepository productRepository) {
         this.orderRepository = orderRepository;
-        this.userRepository = userRepository;
+        this.userFinder = userFinder;
         this.productRepository = productRepository;
     }
 
@@ -37,8 +38,8 @@ public class OrderService {
     public OrderResponse createOrder(OrderRequest request, String phoneNumber) {
 
         //주문한 유저 찾기
-        User user = userRepository.findByPhoneNumber(phoneNumber)
-                .orElseThrow(() -> new CustomException(ErrorMsg.USER_NOT_FOUND));
+        User user = userFinder.findUserByPhoneNumber(phoneNumber);
+
 
         //주문상품 찾기
         Product product = productRepository.findById(request.getProductId())
@@ -53,42 +54,6 @@ public class OrderService {
         //주문 객체 -> DTO로 생성해서 반환
         return new OrderResponse(savedOrders);
 
-    }
-
-    // 사용자가 후기를 작성할 수 있는 주문 목록 조회
-    public List<PendingReviewOrderResponse> getPendingReviewOrdersByUserId(String phoneNumber) {
-
-        //유저 찾기
-        User user = userRepository.findByPhoneNumber(phoneNumber)
-                .orElseThrow(() -> new CustomException(ErrorMsg.USER_NOT_FOUND));
-
-        //후기가 작성되지 않은 주문을 전체조회
-        List<Orders> orders = orderRepository.findAllByUserIdAndReviewedFalse(user.getId());
-
-        //반환해줄 DTO 리스트 타입 선언
-        List<PendingReviewOrderResponse> responseList = new ArrayList<>();
-        for (Orders order : orders) {
-            responseList.add(new PendingReviewOrderResponse(order));
-        }
-        return responseList;
-    }
-
-    // 사용자가 후기를 작성한 주문 목록 조회
-    public List<WrittenReviewOrderResponse> getWrittenReviewOrdersByUserId(String phoneNumber) {
-
-        //유저 찾기
-        User user = userRepository.findByPhoneNumber(phoneNumber)
-                .orElseThrow(() -> new CustomException(ErrorMsg.USER_NOT_FOUND));
-
-        //후기가 작성된 주문 전체조회
-        List<Orders> orders = orderRepository.findAllByUserIdAndReviewedTrue(user.getId());
-
-        //반환해줄 DTO 리스트 타입 선언
-        List<WrittenReviewOrderResponse> responseList = new ArrayList<>();
-        for (Orders order : orders) {
-            responseList.add(new WrittenReviewOrderResponse(order));
-        }
-        return responseList;
     }
 
 }
